Implements the overview page layout
Adds the overview page with balance, income, and expenses cards.

Introduces a new Card component with variants for color and padding to enhance reusability and styling consistency.

Updates the login and signup pages to use the new Card component with padding.

Updates the BasicInput component to support icon buttons with titles, improving accessibility.

Adds disabled styles to Button component